python-service/requirements.txt
astroid==2.3.3
autopep8==1.5
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pika==1.1.0
pycodestyle==2.5.0
pylint==2.4.4
six==1.14.0
wrapt==1.11.2
opentelemetry-api==1.10.0
opentelemetry-exporter-jaeger==1.10.0
opentelemetry-sdk==1.10.0
opentelemetry-exporter-otlp==1.10.0